+++ tests/LinkDotNet.StringBuilder.Benchmarks/AppendValueTypesBenchmark.cs
@@ -6,12 +6,14 @@ namespace LinkDotNet.StringBuilder.Benchmarks;
 [MemoryDiagnoser]
 public class AppendValueTypes
 {
+    private const int NumberOfIterations = 25;
+
     [Benchmark]
     public string DotNetStringBuilder()
     {
         var builder = new System.Text.StringBuilder();
 
-        for (var i = 0; i < 25; i++)
+        for (var i = 0; i < NumberOfIterations; i++)
         {
             builder.Append(true);
             builder.Append(int.MaxValue);
@@ -29,7 +31,7 @@ public class AppendValueTypes
     {
         using var builder = new ValueStringBuilder();
 
-        for (var i = 0; i < 25; i++)
+        for (var i = 0; i < NumberOfIterations; i++)
         {
             builder.Append(true);
             builder.Append(int.MaxValue);
@@ -47,7 +49,7 @@ public class AppendValueTypes
     {
         var builder = ZString.CreateStringBuilder();
 
-        for (var i = 0; i < 25; i++)
+        for (var i = 0; i < NumberOfIterations; i++)
         {
             builder.Append(true);
             builder.Append(int.MaxValue);